XSS: remove forward slash (#516)
Encoding forward slash is an unneeded usage of the encoding. The only
characters that should be encoded would be the reserved characters `&`, `<`,
`>` and `"`:

https://developer.mozilla.org/Glossary/Entity

Further these implementations from major programming language vendors dont
encode forward slash:

- https://docs.microsoft.com/dotnet/api/system.web.httputility.htmlencode
- https://golang.org/pkg/html#EscapeString
- https://php.net/function.htmlentities
- https://docs.python.org/library/html.html#html.escape
- https://ruby-doc.org/stdlib/libdoc/cgi/rdoc/CGI/Util.html#method-i-escapeHTML

I found other implementations as well with same behavior. Note that these
implementations are doing encoding, but just not with the forward slash
character. Further, its not recommended by the HTML5 specification:

https://w3.org/TR/html52/syntax.html#serializing-html-fragments

or even the HTML Living Standard:

https://html.spec.whatwg.org/multipage/parsing.html#serialising-html-fragments